Serialize an R object to JSON

TODO

Usage

serialize_json(
  obj,
  file = NULL,
  collapse = FALSE,
  format = c("pretty", "compact", "oneline")
)

Arguments

obj

R object to serialize.

file

If not NULL then the result if written to this file.

collapse

If file is NULL then whether to return a character scalar or a character vector.

format

Formatting, one of:

"pretty": arrays and objects are formatted in multiple lines,

"compact": format everything without whitespace,

"oneline": format everything without newlines, but include whitespace after commas, colons, opening brackets and braces, and before closing brackets and braces.

Value

If file is NULL then a character scalar (collapse = TRUE) or vector (collapse = FALSE). If file is not NULL then nothing.